Enum known_folders::KnownFolder

#[non_exhaustive]
pub enum KnownFolder {

    AccountPictures,
    AddNewPrograms,
    AdminTools,
    AllAppMods,
    AppCaptures,
    AppDataDesktop,
    AppDataDocuments,
    AppDataFavorites,
    AppDataProgramData,
    AppUpdates,
    ApplicationShortcuts,
    AppsFolder,
    CDBurning,
    CameraRoll,
    CameraRollLibrary,
    ChangeRemovePrograms,
    CommonAdminTools,
    CommonOEMLinks,
    CommonPrograms,
    CommonStartMenu,
    CommonStartMenuPlaces,
    CommonStartup,
    CommonTemplates,
    ComputerFolder,
    ConflictFolder,
    ConnectionsFolder,
    Contacts,
    ControlPanelFolder,
    Cookies,
    CurrentAppMods,
    Desktop,
    DevelopmentFiles,
    Device,
    DeviceMetadataStore,
    Documents,
    DocumentsLibrary,
    Downloads,
    Favorites,
    Fonts,
    GameTasks,
    Games,
    History,
    HomeGroup,
    HomeGroupCurrentUser,
    ImplicitAppShortcuts,
    InternetCache,
    InternetFolder,
    Libraries,
    Links,
    LocalAppData,
    LocalAppDataLow,
    LocalDocuments,
    LocalDownloads,
    LocalMusic,
    LocalPictures,
    LocalStorage,
    LocalVideos,
    LocalizedResourcesDir,
    Music,
    MusicLibrary,
    NetHood,
    NetworkFolder,
    Objects3D,
    OneDrive,
    OriginalImages,
    PhotoAlbums,
    Pictures,
    PicturesLibrary,
    Playlists,
    PrintHood,
    PrintersFolder,
    Profile,
    ProgramData,
    ProgramFiles,
    ProgramFilesCommon,
    ProgramFilesCommonX64,
    ProgramFilesCommonX86,
    ProgramFilesX64,
    ProgramFilesX86,
    Programs,
    Public,
    PublicDesktop,
    PublicDocuments,
    PublicDownloads,
    PublicGameTasks,
    PublicLibraries,
    PublicMusic,
    PublicPictures,
    PublicRingtones,
    PublicUserTiles,
    PublicVideos,
    QuickLaunch,
    Recent,
    RecordedCalls,
    RecordedTVLibrary,
    RecycleBinFolder,
    ResourceDir,
    RetailDemo,
    Ringtones,
    RoamedTileImages,
    RoamingAppData,
    RoamingTiles,
    SEARCH_CSC,
    SEARCH_MAPI,
    SampleMusic,
    SamplePictures,
    SamplePlaylists,
    SampleVideos,
    SavedGames,
    SavedPictures,
    SavedPicturesLibrary,
    SavedSearches,
    Screenshots,
    SearchHistory,
    SearchHome,
    SearchTemplates,
    SendTo,
    SidebarDefaultParts,
    SidebarParts,
    SkyDrive,
    SkyDriveCameraRoll,
    SkyDriveDocuments,
    SkyDriveMusic,
    SkyDrivePictures,
    StartMenu,
    StartMenuAllPrograms,
    Startup,
    SyncManagerFolder,
    SyncResultsFolder,
    SyncSetupFolder,
    System,
    SystemX86,
    Templates,
    UserPinned,
    UserProfiles,
    UserProgramFiles,
    UserProgramFilesCommon,
    UsersFiles,
    UsersLibraries,
    Videos,
    VideosLibrary,
    Windows,

}

GUIDs that identify standard folders registered with the system as Known Folders.

These folders are installed with Windows Vista and later operating systems, and a computer will have only folders appropriate to it installed.

For details on the KNOWNFOLDERID constants this enum represents, please refer to the upstream documentation.

Compatibility Notes

The Known Folders API allows for ISVs to extend the set of Known Folder IDs, but this enum only has support for first-party Known Folder IDs included in [windows_sys].

Examples

use known_folders::{get_known_folder_path, KnownFolder};

let profile_dir = get_known_folder_path(KnownFolder::Profile);

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AccountPictures

Known Folder ID FOLDERID_AccountPictures.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_AccountPictures]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_AccountPictures

AddNewPrograms

Known Folder ID FOLDERID_AddNewPrograms.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_AddNewPrograms]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_AddNewPrograms

AdminTools

Known Folder ID FOLDERID_AdminTools.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_AdminTools]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_AdminTools

AllAppMods

Known Folder ID FOLDERID_AllAppMods.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_AllAppMods]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_AllAppMods

AppCaptures

Known Folder ID FOLDERID_AppCaptures.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_AppCaptures]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_AppCaptures

AppDataDesktop

Known Folder ID FOLDERID_AppDataDesktop.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_AppDataDesktop]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_AppDataDesktop

AppDataDocuments

Known Folder ID FOLDERID_AppDataDocuments.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_AppDataDocuments]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_AppDataDocuments

AppDataFavorites

Known Folder ID FOLDERID_AppDataFavorites.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_AppDataFavorites]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_AppDataFavorites

AppDataProgramData

Known Folder ID FOLDERID_AppDataProgramData.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_AppDataProgramData]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_AppDataProgramData

AppUpdates

Known Folder ID FOLDERID_AppUpdates.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_AppUpdates]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_AppUpdates

ApplicationShortcuts

Known Folder ID FOLDERID_ApplicationShortcuts.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_ApplicationShortcuts]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_ApplicationShortcuts

AppsFolder

Known Folder ID FOLDERID_AppsFolder.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_AppsFolder]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_AppsFolder

CDBurning

Known Folder ID FOLDERID_CDBurning.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_CDBurning]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_CDBurning

CameraRoll

Known Folder ID FOLDERID_CameraRoll.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_CameraRoll]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_CameraRoll

CameraRollLibrary

Known Folder ID FOLDERID_CameraRollLibrary.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_CameraRollLibrary]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_CameraRollLibrary

ChangeRemovePrograms

Known Folder ID FOLDERID_ChangeRemovePrograms.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_ChangeRemovePrograms]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_ChangeRemovePrograms

CommonAdminTools

Known Folder ID FOLDERID_CommonAdminTools.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_CommonAdminTools]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_CommonAdminTools

CommonOEMLinks

Known Folder ID FOLDERID_CommonOEMLinks.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_CommonOEMLinks]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_CommonOEMLinks

CommonPrograms

Known Folder ID FOLDERID_CommonPrograms.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_CommonPrograms]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_CommonPrograms

CommonStartMenu

Known Folder ID FOLDERID_CommonStartMenu.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_CommonStartMenu]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_CommonStartMenu

CommonStartMenuPlaces

Known Folder ID FOLDERID_CommonStartMenuPlaces.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_CommonStartMenuPlaces]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_CommonStartMenuPlaces

CommonStartup

Known Folder ID FOLDERID_CommonStartup.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_CommonStartup]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_CommonStartup

CommonTemplates

Known Folder ID FOLDERID_CommonTemplates.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_CommonTemplates]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_CommonTemplates

ComputerFolder

Known Folder ID FOLDERID_ComputerFolder.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_ComputerFolder]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_ComputerFolder

ConflictFolder

Known Folder ID FOLDERID_ConflictFolder.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_ConflictFolder]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_ConflictFolder

ConnectionsFolder

Known Folder ID FOLDERID_ConnectionsFolder.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_ConnectionsFolder]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_ConnectionsFolder

Contacts

Known Folder ID FOLDERID_Contacts.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Contacts]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Contacts

ControlPanelFolder

Known Folder ID FOLDERID_ControlPanelFolder.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_ControlPanelFolder]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_ControlPanelFolder

Cookies

Known Folder ID FOLDERID_Cookies.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Cookies]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Cookies

CurrentAppMods

Known Folder ID FOLDERID_CurrentAppMods.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_CurrentAppMods]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_CurrentAppMods

Desktop

Known Folder ID FOLDERID_Desktop.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Desktop]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Desktop

DevelopmentFiles

Known Folder ID FOLDERID_DevelopmentFiles.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_DevelopmentFiles]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_DevelopmentFiles

Device

Known Folder ID FOLDERID_Device.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Device]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Device

DeviceMetadataStore

Known Folder ID FOLDERID_DeviceMetadataStore.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_DeviceMetadataStore]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_DeviceMetadataStore

Documents

Known Folder ID FOLDERID_Documents.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Documents]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Documents

DocumentsLibrary

Known Folder ID FOLDERID_DocumentsLibrary.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_DocumentsLibrary]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_DocumentsLibrary

Downloads

Known Folder ID FOLDERID_Downloads.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Downloads]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Downloads

Favorites

Known Folder ID FOLDERID_Favorites.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Favorites]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Favorites

Fonts

Known Folder ID FOLDERID_Fonts.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Fonts]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Fonts

GameTasks

Known Folder ID FOLDERID_GameTasks.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_GameTasks]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_GameTasks

Games

Known Folder ID FOLDERID_Games.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Games]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Games

History

Known Folder ID FOLDERID_History.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_History]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_History

HomeGroup

Known Folder ID FOLDERID_HomeGroup.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_HomeGroup]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_HomeGroup

HomeGroupCurrentUser

Known Folder ID FOLDERID_HomeGroupCurrentUser.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_HomeGroupCurrentUser]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_HomeGroupCurrentUser

ImplicitAppShortcuts

Known Folder ID FOLDERID_ImplicitAppShortcuts.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_ImplicitAppShortcuts]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_ImplicitAppShortcuts

InternetCache

Known Folder ID FOLDERID_InternetCache.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_InternetCache]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_InternetCache

InternetFolder

Known Folder ID FOLDERID_InternetFolder.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_InternetFolder]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_InternetFolder

Libraries

Known Folder ID FOLDERID_Libraries.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Libraries]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Libraries

Links

Known Folder ID FOLDERID_Links.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Links]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Links

LocalAppData

Known Folder ID FOLDERID_LocalAppData.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_LocalAppData]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_LocalAppData

LocalAppDataLow

Known Folder ID FOLDERID_LocalAppDataLow.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_LocalAppDataLow]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_LocalAppDataLow

LocalDocuments

Known Folder ID FOLDERID_LocalDocuments.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_LocalDocuments]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_LocalDocuments

LocalDownloads

Known Folder ID FOLDERID_LocalDownloads.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_LocalDownloads]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_LocalDownloads

LocalMusic

Known Folder ID FOLDERID_LocalMusic.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_LocalMusic]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_LocalMusic

LocalPictures

Known Folder ID FOLDERID_LocalPictures.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_LocalPictures]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_LocalPictures

LocalStorage

Known Folder ID FOLDERID_LocalStorage.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_LocalStorage]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_LocalStorage

LocalVideos

Known Folder ID FOLDERID_LocalVideos.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_LocalVideos]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_LocalVideos

LocalizedResourcesDir

Known Folder ID FOLDERID_LocalizedResourcesDir.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_LocalizedResourcesDir]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_LocalizedResourcesDir

Music

Known Folder ID FOLDERID_Music.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Music]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Music

MusicLibrary

Known Folder ID FOLDERID_MusicLibrary.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_MusicLibrary]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_MusicLibrary

NetHood

Known Folder ID FOLDERID_NetHood.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_NetHood]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_NetHood

NetworkFolder

Known Folder ID FOLDERID_NetworkFolder.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_NetworkFolder]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_NetworkFolder

Objects3D

Known Folder ID FOLDERID_Objects3D.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Objects3D]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Objects3D

OneDrive

Known Folder ID FOLDERID_OneDrive.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_OneDrive]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_OneDrive

OriginalImages

Known Folder ID FOLDERID_OriginalImages.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_OriginalImages]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_OriginalImages

PhotoAlbums

Known Folder ID FOLDERID_PhotoAlbums.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_PhotoAlbums]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_PhotoAlbums

Pictures

Known Folder ID FOLDERID_Pictures.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Pictures]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Pictures

PicturesLibrary

Known Folder ID FOLDERID_PicturesLibrary.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_PicturesLibrary]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_PicturesLibrary

Playlists

Known Folder ID FOLDERID_Playlists.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Playlists]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Playlists

PrintHood

Known Folder ID FOLDERID_PrintHood.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_PrintHood]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_PrintHood

PrintersFolder

Known Folder ID FOLDERID_PrintersFolder.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_PrintersFolder]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_PrintersFolder

Profile

Known Folder ID FOLDERID_Profile.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Profile]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Profile

ProgramData

Known Folder ID FOLDERID_ProgramData.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_ProgramData]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_ProgramData

ProgramFiles

Known Folder ID FOLDERID_ProgramFiles.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_ProgramFiles]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_ProgramFiles

ProgramFilesCommon

Known Folder ID FOLDERID_ProgramFilesCommon.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_ProgramFilesCommon]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_ProgramFilesCommon

ProgramFilesCommonX64

Known Folder ID FOLDERID_ProgramFilesCommonX64.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_ProgramFilesCommonX64]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_ProgramFilesCommonX64

ProgramFilesCommonX86

Known Folder ID FOLDERID_ProgramFilesCommonX86.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_ProgramFilesCommonX86]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_ProgramFilesCommonX86

ProgramFilesX64

Known Folder ID FOLDERID_ProgramFilesX64.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_ProgramFilesX64]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_ProgramFilesX64

ProgramFilesX86

Known Folder ID FOLDERID_ProgramFilesX86.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_ProgramFilesX86]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_ProgramFilesX86

Programs

Known Folder ID FOLDERID_Programs.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Programs]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Programs

Public

Known Folder ID FOLDERID_Public.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Public]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Public

PublicDesktop

Known Folder ID FOLDERID_PublicDesktop.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_PublicDesktop]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_PublicDesktop

PublicDocuments

Known Folder ID FOLDERID_PublicDocuments.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_PublicDocuments]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_PublicDocuments

PublicDownloads

Known Folder ID FOLDERID_PublicDownloads.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_PublicDownloads]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_PublicDownloads

PublicGameTasks

Known Folder ID FOLDERID_PublicGameTasks.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_PublicGameTasks]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_PublicGameTasks

PublicLibraries

Known Folder ID FOLDERID_PublicLibraries.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_PublicLibraries]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_PublicLibraries

PublicMusic

Known Folder ID FOLDERID_PublicMusic.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_PublicMusic]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_PublicMusic

PublicPictures

Known Folder ID FOLDERID_PublicPictures.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_PublicPictures]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_PublicPictures

PublicRingtones

Known Folder ID FOLDERID_PublicRingtones.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_PublicRingtones]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_PublicRingtones

PublicUserTiles

Known Folder ID FOLDERID_PublicUserTiles.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_PublicUserTiles]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_PublicUserTiles

PublicVideos

Known Folder ID FOLDERID_PublicVideos.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_PublicVideos]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_PublicVideos

QuickLaunch

Known Folder ID FOLDERID_QuickLaunch.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_QuickLaunch]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_QuickLaunch

Recent

Known Folder ID FOLDERID_Recent.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Recent]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Recent

RecordedCalls

Known Folder ID FOLDERID_RecordedCalls.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_RecordedCalls]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_RecordedCalls

RecordedTVLibrary

Known Folder ID FOLDERID_RecordedTVLibrary.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_RecordedTVLibrary]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_RecordedTVLibrary

RecycleBinFolder

Known Folder ID FOLDERID_RecycleBinFolder.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_RecycleBinFolder]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_RecycleBinFolder

ResourceDir

Known Folder ID FOLDERID_ResourceDir.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_ResourceDir]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_ResourceDir

RetailDemo

Known Folder ID FOLDERID_RetailDemo.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_RetailDemo]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_RetailDemo

Ringtones

Known Folder ID FOLDERID_Ringtones.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Ringtones]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Ringtones

RoamedTileImages

Known Folder ID FOLDERID_RoamedTileImages.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_RoamedTileImages]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_RoamedTileImages

RoamingAppData

Known Folder ID FOLDERID_RoamingAppData.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_RoamingAppData]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_RoamingAppData

RoamingTiles

Known Folder ID FOLDERID_RoamingTiles.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_RoamingTiles]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_RoamingTiles

SEARCH_CSC

Known Folder ID FOLDERID_SEARCH_CSC.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SEARCH_CSC]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SEARCH_CSC

SEARCH_MAPI

Known Folder ID FOLDERID_SEARCH_MAPI.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SEARCH_MAPI]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SEARCH_MAPI

SampleMusic

Known Folder ID FOLDERID_SampleMusic.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SampleMusic]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SampleMusic

SamplePictures

Known Folder ID FOLDERID_SamplePictures.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SamplePictures]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SamplePictures

SamplePlaylists

Known Folder ID FOLDERID_SamplePlaylists.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SamplePlaylists]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SamplePlaylists

SampleVideos

Known Folder ID FOLDERID_SampleVideos.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SampleVideos]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SampleVideos

SavedGames

Known Folder ID FOLDERID_SavedGames.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SavedGames]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SavedGames

SavedPictures

Known Folder ID FOLDERID_SavedPictures.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SavedPictures]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SavedPictures

SavedPicturesLibrary

Known Folder ID FOLDERID_SavedPicturesLibrary.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SavedPicturesLibrary]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SavedPicturesLibrary

SavedSearches

Known Folder ID FOLDERID_SavedSearches.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SavedSearches]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SavedSearches

Screenshots

Known Folder ID FOLDERID_Screenshots.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Screenshots]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Screenshots

SearchHistory

Known Folder ID FOLDERID_SearchHistory.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SearchHistory]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SearchHistory

SearchHome

Known Folder ID FOLDERID_SearchHome.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SearchHome]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SearchHome

SearchTemplates

Known Folder ID FOLDERID_SearchTemplates.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SearchTemplates]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SearchTemplates

SendTo

Known Folder ID FOLDERID_SendTo.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SendTo]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SendTo

SidebarDefaultParts

Known Folder ID FOLDERID_SidebarDefaultParts.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SidebarDefaultParts]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SidebarDefaultParts

SidebarParts

Known Folder ID FOLDERID_SidebarParts.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SidebarParts]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SidebarParts

SkyDrive

Known Folder ID FOLDERID_SkyDrive.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SkyDrive]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SkyDrive

SkyDriveCameraRoll

Known Folder ID FOLDERID_SkyDriveCameraRoll.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SkyDriveCameraRoll]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SkyDriveCameraRoll

SkyDriveDocuments

Known Folder ID FOLDERID_SkyDriveDocuments.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SkyDriveDocuments]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SkyDriveDocuments

SkyDriveMusic

Known Folder ID FOLDERID_SkyDriveMusic.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SkyDriveMusic]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SkyDriveMusic

SkyDrivePictures

Known Folder ID FOLDERID_SkyDrivePictures.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SkyDrivePictures]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SkyDrivePictures

StartMenu

Known Folder ID FOLDERID_StartMenu.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_StartMenu]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_StartMenu

StartMenuAllPrograms

Known Folder ID FOLDERID_StartMenuAllPrograms.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_StartMenuAllPrograms]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_StartMenuAllPrograms

Startup

Known Folder ID FOLDERID_Startup.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Startup]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Startup

SyncManagerFolder

Known Folder ID FOLDERID_SyncManagerFolder.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SyncManagerFolder]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SyncManagerFolder

SyncResultsFolder

Known Folder ID FOLDERID_SyncResultsFolder.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SyncResultsFolder]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SyncResultsFolder

SyncSetupFolder

Known Folder ID FOLDERID_SyncSetupFolder.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SyncSetupFolder]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SyncSetupFolder

System

Known Folder ID FOLDERID_System.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_System]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_System

SystemX86

Known Folder ID FOLDERID_SystemX86.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_SystemX86]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_SystemX86

Templates

Known Folder ID FOLDERID_Templates.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Templates]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Templates

UserPinned

Known Folder ID FOLDERID_UserPinned.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_UserPinned]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_UserPinned

UserProfiles

Known Folder ID FOLDERID_UserProfiles.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_UserProfiles]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_UserProfiles

UserProgramFiles

Known Folder ID FOLDERID_UserProgramFiles.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_UserProgramFiles]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_UserProgramFiles

UserProgramFilesCommon

Known Folder ID FOLDERID_UserProgramFilesCommon.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_UserProgramFilesCommon]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_UserProgramFilesCommon

UsersFiles

Known Folder ID FOLDERID_UsersFiles.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_UsersFiles]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_UsersFiles

UsersLibraries

Known Folder ID FOLDERID_UsersLibraries.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_UsersLibraries]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_UsersLibraries

Videos

Known Folder ID FOLDERID_Videos.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Videos]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Videos

VideosLibrary

Known Folder ID FOLDERID_VideosLibrary.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_VideosLibrary]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_VideosLibrary

Windows

Known Folder ID FOLDERID_Windows.

Upstream Documentation

[windows_sys::Win32::UI::Shell::FOLDERID_Windows]

https://learn.microsoft.com/en-us/windows/win32/shell/knownfolderid#FOLDERID_Windows